Move contenteditable directive to separate file
